requirements.txt
streamlit>=1.22.0
pillow>=9.4.0
requests>=2.28.2
numpy>=1.23.5 